databaseをserverに合わせた

+++ Mobile/app/src/main/java/com/example/hideandseek/data/datasource/remote/PostData.kt
@@ -5,9 +5,9 @@ import com.squareup.moshi.Json
 class PostData {
     data class PostSpacetime(
         @Json(name = "time") val time: String,
-        @Json(name = "latitude") val latitude: Float,
-        @Json(name = "longitude") val longitude: Float,
-        @Json(name = "altitude") val altitude: Float,
+        @Json(name = "latitude") val latitude: Double,
+        @Json(name = "longitude") val longitude: Double,
+        @Json(name = "altitude") val altitude: Double,
         @Json(name = "objId") val objId: Int
     )
 }
